kaeru: Support both `kobe2` variants
There appear to be two different variants of this device: one with LTE and
another with Wi-Fi only. Their firmware versions differ significantly, and
crossflashing is not possible, as it breaks features such as Wi-Fi or RIL.

To resolve this, add separate support for both variants by using different
codenames and using different offsets accordingly.

Support for the LTE variant has been tested by @bluuesmokes.

Change-Id: I9ba2a0787b2edfdb747724ac8e56bfd606946985